backend/requirements.txt
django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg[binary]==3.2.10
drf-spectacular==0.29.0
